[PW-7096] - Remove NotificationHandler and move methods to NotificationService
