+++ src/Service/NotificationService.php
@@ -24,6 +24,7 @@
 
 namespace Adyen\Shopware\Service;
 
+use Adyen\Shopware\Entity\Notification\NotificationEntity;
 use Adyen\Shopware\ScheduledTask\ProcessNotificationsHandler;
 use DateTimeInterface;
 use Shopware\Core\Framework\Context;
@@ -243,5 +244,51 @@ class NotificationService
             ],
             Context::createDefaultContext()
         );
+    }
+
+    /**
+     * Calculates the scheduled processing time according to notification type.
+     *
+     * @param NotificationEntity $notification
+     * @param bool $reschedule
+     * @return \DateTimeInterface|null
+     */
+    public function calculateScheduledProcessingTime(NotificationEntity $notification, bool $reschedule = false)
+    {
+        if ($reschedule) {
+            $scheduledProcessingTime = new \DateTime();
+        } else {
+            $scheduledProcessingTime = $notification->getCreatedAt();
+        }
+
+        switch ($notification->getEventCode()) {
+            case 'AUTHORISATION':
+                if (!$notification->isSuccess()) {
+                    $scheduledProcessingTime = $scheduledProcessingTime->add(new \DateInterval('PT30M'));
+                }
+                break;
+            case 'OFFER_CLOSED':
+                $scheduledProcessingTime = $scheduledProcessingTime->add(new \DateInterval('PT30M'));
+                break;
+            default:
+                break;
+        }
+
+        return $scheduledProcessingTime;
+    }
+
+    /**
+     * @param NotificationEntity $notification
+     * @return bool
+     */
+    public function canBeRescheduled(NotificationEntity $notification): bool
+    {
+        $timeDifferenceInDays = $notification->getScheduledProcessingTime()->diff(new \DateTime())->format('%a');
+
+        if (!$notification->isDone() && $timeDifferenceInDays >= 1) {
+            return true;
+        } else {
+            return false;
+        }
     }
 }
